Updated Processing code and EMGServoTest
Added the ability to communicate with the Arduino through the use of
simple commands. Changed GUI to have a text section where text and
various states (if there are any) will be displayed. Thresholds are now
drawn as a single line when serial data transmission starts.

In EMGServoTest, added new commands PEAC and ROCK (Two new hand symbols
for testing) and added else if to a lot of the if statements. Changed
the exit condition for the fingerControl method to include any value
not in the valid grouping of numbers. Might change that back…