Refined processing code and multiServoTest code
General improvements to processing code -
Displays useful data in the text data section, added a method to draw
the text, and a method to wipe it again (have to draw a rectangle over
the text, then draw the text again to update it…) Also, we now draw the
thresholds in setup() AND when the graph starts to go offscreen. This
should prevent it from being extremely buggy with the out of place
diagonal lines everywhere. The improvements can be summed up as things
only taking place when necessary. (Thresholds used to be drawn every
loop…)

In multiEMGServoTest, implemented averaging and enabled all three EMG
readers. Added upper thresholds for each EMG, and briefly tested all of
the code.