Added flags for multiEMGServoTest activations, fixed processing bugs.
All EMGs now activate/deactivate %something% based on the flag. More
flags can be created for actions to be tracked separately. Fixed a bug
in processing where the thresholds were not redrawn after the graph
updated (due to the order in which things were drawn — first
thresholds, then the screen was reset, which reset the thresholds that
were just drawn.)