Changed a few variables in multiEMGServoTest. Fixed a graphing bug in Processing.
There was a big bug in the processing code where graphing did not work
as intended. In order to “connect the dots” on the graph, the change in
value is calculated, and a line with a length equal to the change is
drawn. This relies on knowing the previous value, which was never
actually set. I added code to set the previous value during each loop,
so that the subsequent loop could draw the connecting lines.