Fixed small bugs with processing code & added a feature
Fixed a bug where when drawing rectangles the previously selected
stroke colour would be used on the border. This resulted in the black
rectangle having a blue border, which was just a bit annoying.

Added the “mark” command, which adds a marker in the text file and
draws a vertical line on the graph.