Fixed an issue with EMG averaging and added a new feature to HDP that will draw the graph in a different color when the threshold is active.
With averaging, the average value was ONLY reset following a successful
activation. This caused the averaged value for future activations to be
inaccurate, because it was using part of the previous averaged value.

Processing now draws values that are past the threshold in a brighter
color. In communications, A<value> = below threshold, while a<value> =
greater than or equal to threshold. Example:

A12
B35
C0
a84
B35
C0